:art: start split of code into domains
