+++ src/main/java/agency/place/PlaceModelEventsSubscriber.java
@@ -1,4 +1,4 @@
-package agency.model;
+package agency.place;
 
 public interface PlaceModelEventsSubscriber {
     void subscribe(PlaceModelEvents listener);
