+++ src/main/java/agency/trip/TripModelEvents.java
@@ -1,6 +1,4 @@
-package agency.model;
-
-import agency.data.Trip;
+package agency.trip;
 
 public interface TripModelEvents {
     void onTripCreated(Trip trip);
