+++ src/main/java/agency/trip/TripModelEventsSubscriber.java
@@ -1,4 +1,4 @@
-package agency.model;
+package agency.trip;
 
 public interface TripModelEventsSubscriber {
     void subscribe(TripModelEvents listener);
